Improved: removed unused locale text and removed grandTotal dependency from the items in orders (#1302)
- Changed the fetchProductAverageCost service to fetch records for one product at a time, since we do not add products in bulk.
