+++ src/services/ProductService.ts
@@ -1,6 +1,7 @@
 import { api, apiClient } from '@/adapter';
 import store from '@/store';
-
+import logger from '@/logger';
+import { hasError } from '@/adapter';
 
 const fetchProducts = async (query: any): Promise <any>  => {
   return api({
@@ -28,57 +29,51 @@ const fetchProductComponents = async (payload: any): Promise<any> => {
   });
 }
 
-const fetchProductsAverageCost = async (productIds: any, facilityId: any): Promise<any> => {
+const fetchProductAverageCost = async (productId: string, facilityId: string): Promise<any> => {
   const omstoken = store.getters['user/getUserToken'];
   const baseURL = store.getters['user/getMaargBaseUrl'];
-  
-  if(!productIds.length) return []
-  const requests = [], productIdList = productIds, productAverageCostDetail = {} as any;
 
-  while(productIdList.length) {
-    const productIds = productIdList.splice(0, 100)
-    const params = {
-      customParametersMap: {
-        facilityId,
-        productId: productIds,
-        productId_op: "in",
-        orderByField: "-fromDate",
-        pageIndex: 0,
-        pageSize: 100 //There should be more than one active record per product
+  if(!productId) return;
+  let productAverageCost = ''
+
+  const payload = {
+    customParametersMap: {
+      facilityId,
+      productId,
+      orderByField: "-fromDate",
+      pageIndex: 0,
+      pageSize: 1
+    },
+    dataDocumentId: "ProductWeightedAverageCost",
+    filterByDate: true
+  };
+
+  try {
+    const resp = await apiClient({
+      url: `/oms/dataDocumentView`,
+      method: "post",
+      baseURL,
+      headers: {
+        "Authorization": `Bearer ${omstoken}`,
+        "Content-Type": "application/json"
       },
-      dataDocumentId: "ProductWeightedAverageCost",
-      filterByDate: true
+      data: payload
+    });
+
+    if(!hasError(resp) && resp.data?.entityValueList?.length) {
+      const list = resp.data.entityValueList;
+      productAverageCost = list[0].averageCost;
     }
-    requests.push(params)
+  } catch (err) {
+    logger.error("Failed to fetch product average cost", err);
+    return;
   }
 
-  const productAverageCostResps = await Promise.allSettled(requests.map((payload) => apiClient({
-    url: `/oms/dataDocumentView`,
-    method: "post",
-    baseURL,
-    headers: {
-      "Authorization": "Bearer " + omstoken,
-      "Content-Type": "application/json"
-    },
-    data: payload,
-  })))
-
-  const hasFailedResponse = productAverageCostResps.some((response: any) => response.status !== "fulfilled")
-  if(hasFailedResponse) return {};
-
-  productAverageCostResps.map((response: any) => {
-    if(response.value.data?.entityValueList?.length) {
-      response.value.data.entityValueList.map((item: any) => {
-        if(!productAverageCostDetail[item.productId]) productAverageCostDetail[item.productId] = item.averageCost
-      })
-    }
-  })
-
-  return productAverageCostDetail;
-}
+  return productAverageCost;
+};
 
 export const ProductService = {
   fetchProducts,
   fetchProductComponents,
-  fetchProductsAverageCost
+  fetchProductAverageCost
 }
